header.c mbx_unixscan.c include/private/message0.h
I'm an idiot, with dup() the file table is share by the two
file descriptors,  So when one do mailbox_get_message() it
was influencing the parsing since the offset was changing
the world underneath the parsing code ... duh !! solution
reopen the mailbox readonly.  Thanks dave for pointing this out.