Fix quoted-printable and Q encodings.
The Q encoding in RFC 2047 is not exactly the same as Quoted-Printable
encoding: it differs in its handling of space characters (ASCII 32),
which are encoded as underscores.

* libmailutils/filter/qpflt.c (_qp_decoder): Use xd parameter
to determine whether to treat underscores specially.
(_qp_encoder): Likewise.
(_qp_filter,_Q_filter): Allocate and initialize filter data
accordingly to the kind of encoding.