ali: executable file inclusion.
The "< file" construct tests if file is executable and begins with
the shell interpreter magic (#!/ or #! /).  If so, this file is
executed and its output is collected.  This provides a way for
building aliases on the fly.

* mh/mh_alias.l (buffer_ctx) <exec_p>: New member.
(exec_p): New static variable.
(push_source): If the file is executable and begins with
the "#!/" magic, run it and collect its output.
(pop_source): Use exec_p to decide how to close the file.
* mh/tests/ali.at: Test executable file inclusion.
* NEWS: Update.