maildir: Keep trailing comma even when no flags are set
Without this, `mk_info_filename' generates filenames of the following
form when the flag set is empty:

    1402217787.M664911P5957Q143.hostname:2

The original Maildir spec (http://cr.yp.to/proto/maildir.html),
however, mandates that filenames must be of the form `uniq:info', with
`info' starting with "2,"--including a comma.

This patch removes the "cleanup" performed by `mk_info_filename' as
that there is no rationale for it in mailutils' source code history,
and as it breaks Gnus' "nnmaildir" backend (Cf.  `nnmaildir--add-flag').

* libproto/maildir/mbox.c (mk_info_filename): Keep comma at the end of
filename when the flags suffix is empty.