Vacation: improve compatibility with the Sendmail utility.
* libmu_sieve/extensions/vacation.c (match_addresses): Change arguments:
take primary email address as well as a list of aliases.
(vacation_reply): Reply if the primary email or one of alias emails are
listed in recipient list of the original message.  New tag "always_reply"
reverts to the old behavior: reply regardless of whether the message is
destined for us.
Use the current user email as the sender address of the vacation message.
New tag "return_address" sets the sender address explicitly.
* sieve/tests/vacation.at: Add new tests.